+++ actions/teams.ts
@@ -2,22 +2,31 @@
 
 import prisma from "@/lib/db";
 import fs from "fs/promises";
+import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
 export async function addTeam(prevState: unknown, formData: FormData) {
   const data = Object.fromEntries(formData.entries());
-  const imagePath = `/teams/${crypto.randomUUID()}-${data.image.name}`;
 
-  await fs.writeFile(
-    `public${imagePath}`,
-    Buffer.from(await data.image.arrayBuffer())
-  );
+  let imagePath = "";
+  if (data.image !== null && data.image.size > 0) {
+    const imagePath = `/teams/${crypto.randomUUID()}-${data.image.name}`;
+
+    await fs.writeFile(
+      `public${imagePath}`,
+      Buffer.from(await data.image.arrayBuffer())
+    );
+  }
 
   await prisma.team.create({
     data: {
-      name: data.name,
-      flagUrl: imagePath,
+      name: data.name.toString(),
+      flagUrl: data.image.size > 0 ? imagePath : null,
     },
   });
+
+  revalidatePath("/dashboard/teams");
+  redirect("/dashboard/teams");
 }
 
 export async function updateTeam(
@@ -32,9 +41,11 @@ export async function updateTeam(
   if (team == null) return;
 
   let imagePath = team.flagUrl;
-  if (data.image != null && data.image.size > 0) {
-    await fs.unlink(`public${team.flagUrl}`);
+  if (data.image !== null && data.image.size > 0) {
+    if (team.flagUrl) await fs.unlink(`public${team.flagUrl}`);
+
     imagePath = `/teams/${crypto.randomUUID()}-${data.image.name}`;
+
     await fs.writeFile(
       `public${imagePath}`,
       Buffer.from(await data.image.arrayBuffer())
@@ -44,8 +55,11 @@ export async function updateTeam(
   await prisma.team.update({
     where: { id },
     data: {
-      name: data.name,
+      name: data.name.toString(),
       flagUrl: imagePath,
     },
   });
+
+  revalidatePath("/dashboard/teams");
+  redirect("/dashboard/teams");
 }
